Collapse landmark generation into indexes only
