+++ Tests/Pathfinding/Schemas/testLandmarksExtremes.py
@@ -133,13 +133,13 @@ class testLandmarksExtremes(baseTest):
             {0: 5},
             {0: 428},
         ]
-        actual, _ = foo.get_landmarks(index=True)
+        actual, _ = foo.get_landmarks()
         self.assertEqual(expected, actual)
 
         actual, _ = foo.get_landmarks()
         self.assertEqual(6, len(actual))
         chunk = actual[3]
-        self.assertEqual(0, chunk[0].index)
+        self.assertEqual(0, chunk[0])
 
     def test_component_landmarks_of_zarushagar(self) -> None:
         sourcefile = self.unpack_filename('../DeltaFiles/Zarushagar.sec')
@@ -165,7 +165,7 @@ class testLandmarksExtremes(baseTest):
         expected[6] = {418, 420}
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        _, actual = foo.get_landmarks(index=True)
+        _, actual = foo.get_landmarks()
         self.assertEqual(expected, actual)
 
     def test_landmarks_of_dagudashaag(self) -> None:
@@ -193,7 +193,7 @@ class testLandmarksExtremes(baseTest):
         expected_landmarks = [{0: 555, 2: 426}, {0: 558, 2: 423}, {0: 129}, {0: 138}, {0: 6}, {0: 427}]
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        landmarks, actual = foo.get_landmarks(index=True)
+        landmarks, actual = foo.get_landmarks()
         self.assertEqual(expected_components, actual)
         self.assertEqual(expected_landmarks, landmarks)
 
@@ -220,3 +220,52 @@ class testLandmarksExtremes(baseTest):
 
             foo = LandmarksTriaxialExtremes(galaxy)
             self.assertEqual(expected, foo.max_slots)
+
+    def test_landmarks_on_ibara_subsector_single_component(self) -> None:
+        sourcefile = self.unpack_filename('DeltaFiles/Zarushagar-Ibara.sec')
+
+        sector = SectorDictionary.load_traveller_map_file(sourcefile)
+        delta = DeltaDictionary()
+        delta[sector.name] = sector
+
+        args = self._make_args()
+
+        galaxy = DeltaGalaxy(args.btn, args.max_jump)
+        galaxy.read_sectors(delta, args.pop_code, args.ru_calc,
+                            args.route_reuse, args.routes, args.route_btn, args.mp_threads, args.debug_flag)
+        galaxy.output_path = args.output
+
+        galaxy.generate_routes()
+        galaxy.trade.calculate_components()
+
+        self.assertEqual(1, len(galaxy.trade.components), "Unexpected number of components at J-4")
+
+        expected_landmarks = [{0: 34}, {0: 30}, {0: 27}, {0: 0}]
+        landmarks, _ = galaxy.trade.get_landmarks()
+        self.assertTrue(isinstance(landmarks, list), 'Landmarks result should be a list')
+        self.assertEqual(expected_landmarks, landmarks, 'Unexpected landmark result')
+
+    def test_landmarks_on_ibara_subsector_multiple_components(self) -> None:
+        sourcefile = self.unpack_filename('DeltaFiles/Zarushagar-Ibara.sec')
+
+        sector = SectorDictionary.load_traveller_map_file(sourcefile)
+        delta = DeltaDictionary()
+        delta[sector.name] = sector
+
+        args = self._make_args()
+        args.max_jump = 1
+
+        galaxy = DeltaGalaxy(args.btn, args.max_jump)
+        galaxy.read_sectors(delta, args.pop_code, args.ru_calc,
+                            args.route_reuse, args.routes, args.route_btn, args.mp_threads, args.debug_flag)
+        galaxy.output_path = args.output
+
+        galaxy.generate_routes()
+        galaxy.trade.calculate_components()
+
+        self.assertEqual(6, len(galaxy.trade.components), "Unexpected number of components at J-1")
+
+        expected_landmarks = [{0: 29, 2: 13, 4: 34}, {0: 26, 2: 4, 4: 36}, {0: 19}]
+        landmarks, _ = galaxy.trade.get_landmarks()
+        self.assertEqual(3, len(landmarks), 'Should have one landmark per component')
+        self.assertEqual(expected_landmarks, landmarks, 'Unexpected landmark result')
