+++ Tests/Pathfinding/Schemas/testLandmarksExtremesRegression.py
@@ -39,5 +39,5 @@ class testLandmarksExtremesRegression(baseTest):
         expected = [
             {0: 56}, {0: 95}, {0: 43}, {0: 59}, {0: 0}
         ]
-        actual, _ = foo.get_landmarks(index=True, btn=btn)
+        actual, _ = foo.get_landmarks(btn=btn)
         self.assertEqual(expected, actual)
