+++ Tests/Pathfinding/testApproximateShortestPathForest.py
@@ -33,7 +33,7 @@ class testApproximateShortestPathForest(baseTest):
         galaxy = self.set_up_zarushagar_sector()
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        landmarks, _ = foo.get_landmarks(index=True)
+        landmarks, _ = foo.get_landmarks()
         graph = galaxy.stars
         stars = list(graph.nodes)
         source = stars[0]
@@ -54,7 +54,7 @@ class testApproximateShortestPathForest(baseTest):
         galaxy = self.set_up_zarushagar_sector()
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        landmarks, _ = foo.get_landmarks(index=True)
+        landmarks, _ = foo.get_landmarks()
         graph = galaxy.stars
         stars = list(graph.nodes)
         source = stars[0]
@@ -66,7 +66,7 @@ class testApproximateShortestPathForest(baseTest):
         galaxy = self.set_up_zarushagar_sector()
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        landmarks, _ = foo.get_landmarks(index=True)
+        landmarks, _ = foo.get_landmarks()
         graph = galaxy.stars
         source = [item for item in graph if graph.nodes()[item]['star'].component == 0][0]
         targ = [item for item in graph if graph.nodes()[item]['star'].component == 1][0]
@@ -81,7 +81,7 @@ class testApproximateShortestPathForest(baseTest):
         galaxy = self.set_up_zarushagar_sector()
 
         foo = LandmarksTriaxialExtremes(galaxy)
-        landmarks, _ = foo.get_landmarks(index=True)
+        landmarks, _ = foo.get_landmarks()
         graph = galaxy.stars
         stars = list(graph.nodes)
         source = stars[0]
